deploy_package/tp_education_system/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-multipart
pydantic-settings
pandas
openpyxl
apscheduler
python-dateutil
beautifulsoup4
lxml
python-docx
chardet
